backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.0

# Data validation
pydantic==2.8.2
pydantic-settings==2.1.0
email-validator==2.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
clerk-backend-api==1.0.0

# Payments
stripe>=8.0.0

# HTTP requests
requests==2.32.3
httpx==0.27.0

# Utilities
numpy==2.1.1  # DEV-018 deal matching scoring
python-dotenv==1.0.0

# Task Queue (for future async job processing)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

openai>=1.3.0
anthropic>=0.18.0  # Claude AI for PMI reasoning features

# Storage (S3/R2)
boto3>=1.28.0
botocore>=1.31.0

# Video/Audio Processing
ffmpeg-python>=0.2.0
Pillow>=10.0.0
pydub>=0.25.1

# Document Generation
reportlab>=4.0.0  # PDF generation
python-docx>=1.1.0  # DOCX generation
weasyprint>=60.0  # HTML to PDF (alternative)

# Data Processing & Excel
pandas>=2.0.0  # Data processing and Excel export
openpyxl>=3.1.0  # Excel file generation

# Accounting Platform Integrations
xero-python>=5.0.0  # Xero accounting API (Phase 3)
python-quickbooks>=0.9.5  # QuickBooks Online API (Phase 4)
intuit-oauth>=1.2.4  # QuickBooks OAuth 2.0 authentication (Phase 4)
# NetSuite (Phase 11) - uses requests library (already included above) for REST API
